Adding some startup delays to guarrantee the join state.
When the nodes join to the cluster for the first time, only one node at
a time can be at join state. For this reason, we have to make sure that
one node at a time joins the cluster. That's why a delay was added per
module, and with conjuction with an internal delay prior starting each
instance, this should guarrantee (in theory) the join restriction.